fix: delete user 1.0.0

+++ src/usecases/users/delete/delete-user-usecase.ts
@@ -1,4 +1,5 @@
 import { IUserRelations } from "@/dtos/user-relations.dto";
+import { IOrderRepository } from "@/repositories/interfaces/interface-order-repository";
 import { IShoppingCartRepository } from "@/repositories/interfaces/interface-shopping-cart-repository";
 import { IUsersRepository } from "@/repositories/interfaces/interface-users-repository";
 import { AppError } from "@/usecases/errors/app-error";
@@ -13,7 +14,7 @@ interface IRequestDeleteUser {
 export class DeleteUserUseCase{
     constructor(
         private usersRepository: IUsersRepository,
-        private shoppingCartsRepository: IShoppingCartRepository
+        private orderRepository: IOrderRepository
     ) {}
 
     async execute({
@@ -32,6 +33,12 @@ export class DeleteUserUseCase{
 
         if(!passwordMatch){
             throw new AppError('Password not match', 401)
+        }
+
+        const hasOpenOrder = await this.orderRepository.hasActiveOrder(findUserExist.id)
+
+        if(hasOpenOrder){
+            throw new AppError('Existem pedidos viculados ao usuário', 400)
         }
 
         await this.usersRepository.delete(findUserExist.id)
